backend/requirements.txt
# FastAPI core
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Pydantic
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0.post1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9

# Environment
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
